Implement todo ownership, authorized access and more

+++ app/main.py
@@ -1,3 +1,4 @@
+from sqlite3 import IntegrityError
 import sqlalchemy.exc
 from sqlalchemy.orm import Session
 from fastapi import Depends, Response, HTTPException, Request
@@ -6,7 +7,7 @@ from app import api
 from app.auth import authenticate_user_password, verify_auth_token
 from app.auth import generate_new_token_pair
 from app.auth import register_user
-from app.schemas import TodoSchema, TodoCreateSchema, TodoUpdateSchema
+from app.schemas import TodoSchema, TodoCreateSchema, TodoUpdateSchema, UserPermissionsEnum
 from app.schemas import UserSchema
 from app.schemas import UserDTOSchema, TokenPair, TokenRefreshRequest
 from app.db import SessionLocal, Base, engine
@@ -36,23 +37,42 @@ def http_exception_handler(request: Request, exc: HTTPException):
 @api.get('/api/todos', response_model=list[TodoSchema])
 def get_todos_list(offset: int = 0,
                    limit: int = 100,
-                   db: Session = Depends(get_db)):
-    result = crud.get_todos(db, offset, limit)
+                   db: Session = Depends(get_db),
+                   token: dict[str, any] = Depends(verify_auth_token)):
+
+    print(f"getting for user {token['sub']}")
+    permissions: list[str] = token['permissions']
+    if UserPermissionsEnum.personal_read not in permissions:
+        raise HTTPException(status_code=403, detail="Not authorized")
+
+    result = crud.get_todos_for_user(db, offset, limit, token['sub'])
     return result
 
 
 @api.post('/api/todos', response_model=TodoSchema, status_code=201)
 def create_todo(todo: TodoCreateSchema,
                 response: Response,
-                db: Session = Depends(get_db)):
-    result = crud.create_todo(db, todo)
-    resource_url = api.url_path_for('get_todo_by_id', todo=result.id)
+                db: Session = Depends(get_db),
+                token: dict[str, any] = Depends(verify_auth_token)):
+    permissions = token['permission']
+    if UserPermissionsEnum.personal_write not in permissions:
+        raise HTTPException(status_code=403, detail="Not authorized")
+    try:
+        result = crud.create_todo(db, todo, token['sub'])
+    except IntegrityError:
+        raise HTTPException(status_code=400, detail=f"Todo Name Exists in the user scope")
+    resource_url = api.url_path_for('get_todo_by_id', todo_id=result.id)
     response.headers['Location'] = resource_url
     return result
 
 
 @api.get('/api/todos/{todo_id}', response_model=TodoSchema)
-def get_todo_by_id(todo_id: int, db: Session = Depends(get_db)):
+def get_todo_by_id(todo_id: int,
+                   db: Session = Depends(get_db),
+                   token: dict[str, any] = Depends(verify_auth_token)):
+    permissions = token['permissions']
+    if UserPermissionsEnum.personal_read not in permissions:
+        raise HTTPException(status_code=403, detail="Not authorized")
     result = crud.get_todo_by_id(db, todo_id)
     return result
 
@@ -60,7 +80,13 @@ def get_todo_by_id(todo_id: int, db: Session = Depends(get_db)):
 @api.put('/api/todos/{todo_id}', response_model=TodoSchema)
 def update_todo_by_id(todo_id: int,
                       update: TodoUpdateSchema,
-                      db: Session = Depends(get_db)):
+                      db: Session = Depends(get_db),
+                      token: dict[str, any] = Depends(verify_auth_token)):
+    permissions = token['permissions']
+    if UserPermissionsEnum.personal_read not in permissions or\
+       UserPermissionsEnum.personal_write not in permissions:
+        raise HTTPException(status_code=403, detail="Not authorized")
+
     result = crud.update_todo_by_id(db, todo_id, update)
     print(result.name, result.description)
     return result
@@ -72,6 +98,11 @@ def delete_todo_by_id(
     response: Response,
     db: Session = Depends(get_db),
 ):
+    # TODO check user write permissions
+    permissions = token['permissions']
+    if UserPermissionsEnum.personal_read not in permissions or\
+       UserPermissionsEnum.personal_write not in permissions:
+        raise HTTPException(status_code=403, detail="Not authorized")
     if crud.delete_todo_by_id(db, todo_id) == 0:
         raise HTTPException(404, 'Not Found')
